Enable screenshots of animated components
This adds support for capturing animated PNGs of components annotated with `@ShowkaseComposable`, by specifying the new `screenshotConfig` parameter. For example, to capture an animation for 2 seconds with a framerate of 60fps:
```kotlin
@ShowkaseComposable(
    name = "MyAnimatedComponent",
    group = "AnimatedGroup",
    defaultStyle = true,
    screenshotCaptureType = ScreenshotCaptureType.SingleAnimatedImage,
    captureDurationMillis = 2000,
    captureFramerate = 60,
)
@Composable
fun MyAnimatedComponentPreview_Default() {
    var animatedComponentState by remember { mutableStateOf(...) }

    LaunchedEffect(Unit) {
        animatedComponentState = ...
    }

    MyAnimatedComponent(animatedComponentState)
}

```

Updates Paparazzi to `2.0.0-alpha02` to fix [this issue](https://github.com/cashapp/paparazzi/issues/1877) with compileSdk 36.